new file and update struct
